dding functional tests for MYA status page
