+++ test/browser/6-mya.test.ts
@@ -5,8 +5,10 @@ import { LoginPage } from 'test/page-objects/login';
 import { TaskListPage } from 'test/page-objects/task-list';
 import { AssignCasePage } from 'test/page-objects/assign-case';
 import { StatusPage } from 'test/page-objects/status';
+import { oralAppealStages } from 'app/server/data/appealStages';
+const i18n = require('locale/en');
 
-describe('Manage your appeal app @mya', () => {
+describe('Manage your appeal app @mya using COR case', () => {
   let ccdCase;
   let page: Page;
   let loginPage: LoginPage;
@@ -24,7 +26,7 @@ describe('Manage your appeal app @mya', () => {
     await taskListPage.setCookie('manageYourAppeal', 'true');
 
     await loginPage.visitPage(`?tya=${appellantTya}`);
-    await loginPage.login(sidamUser.email || '', sidamUser.password || '');
+    await loginPage.login(sidamUser.email || 'oral.hearingBooked@example.com', sidamUser.password || '');
   });
 
   after(async () => {
@@ -43,10 +45,65 @@ describe('Manage your appeal app @mya', () => {
 
     statusPage.verifyPage();
   });
-  it('signs out and prevents access to pages', async () => {
-    await statusPage.signOut();
-    loginPage.verifyPage();
-    await statusPage.visitPage();
-    loginPage.verifyPage();
+
+  describe('Status page', () => {
+    it('should display navigation tabs and Status tab should be active', async() => {
+      statusPage.verifyPage();
+      expect(await statusPage.getElementText('.navigation-tabs')).to.not.be.null;
+      expect(await statusPage.getElementText('.navigation-tabs ul li.selected')).contain(i18n.statusTab.tabHeader);
+    });
+
+    it('should display subheading', async() => {
+      statusPage.verifyPage();
+      expect(await statusPage.getElementText('.task-list h2')).to.equal(i18n.statusTab.header);
+    });
+
+    it('should display status bar', async() => {
+      statusPage.verifyPage();
+      expect(await statusPage.getElementText('.task-list h2')).to.equal(i18n.statusTab.header);
+    });
+
+    it('should display active stage in status bar for Oral case in hearing booked status', async() => {
+      statusPage.verifyPage();
+      const activeStages = await statusPage.getElementsText('.stage--active');
+      expect(activeStages[activeStages.length - 1]).to.equal(oralAppealStages[4].title);
+    });
+
+    it('should display panel with latest update', async() => {
+      statusPage.verifyPage();
+      expect(await statusPage.getElementText('.panel')).contain(i18n.statusTab.panelHeader);
+    });
+
+    it('should display Help and Support links', async() => {
+      statusPage.verifyPage();
+      expect(await statusPage.getElementText('.mya-contact__content h2')).to.equal(i18n.helpGuides.header);
+      expect(await statusPage.getElementText('.mya-contact__content .govuk-list')).contain(i18n.helpGuides.representatives.linkHeader);
+      expect(await statusPage.getElementText('.mya-contact__content .govuk-list')).contain(i18n.helpGuides.withdrawAppeal.linkHeader);
+    });
+
+    it('should display Contact us for help options and open details', async() => {
+      statusPage.verifyPage();
+      expect(await statusPage.getElementText('.govuk-details.contact-us')).to.equal(i18n.contactUs.title);
+      const elementHandle = await page.$('.govuk-details.contact-us');
+      const heightClosed = await page.evaluate(element => {
+        const { height } = element.getBoundingClientRect();
+        return height;
+      }, elementHandle);
+
+      expect(heightClosed).to.equal(40);
+    });
+
+    it('should open Contact us details', async() => {
+      statusPage.verifyPage();
+      const elementHandle = await page.$('.govuk-details.contact-us');
+      await statusPage.openDetails('.govuk-details.contact-us');
+      const heightOpen = await page.evaluate(element => {
+        const { height } = element.getBoundingClientRect();
+        return height;
+      }, elementHandle);
+
+      expect(heightOpen).to.equal(385);
+    });
   });
+
 });
