+++ views/components/mya-contact-us.html
@@ -6,10 +6,10 @@
       <div class="govuk-grid-column-two-thirds">
         <div class="mya-contact__content">
           {% if showSupport %}
-            <h2 class="govuk-heading-s govuk-!-margin-bottom-1">Help and support</h2>
+            <h2 class="govuk-heading-s govuk-!-margin-bottom-1">{{ i18n.helpGuides.header }}</h2>
             <ul class="govuk-list">
-              <li><a href="/support-representatives">Representatives</a></li>
-              <li><a href="/support-withdraw-appeal">Withdrawing an appeal</a></li>
+              <li><a href="/support-representatives">{{ i18n.helpGuides.representatives.linkHeader }}</a></li>
+              <li><a href="/support-withdraw-appeal">{{ i18n.helpGuides.withdrawAppeal.linkHeader }}</a></li>
             </ul>
           {% endif %}
           <details class="govuk-details contact-us">
