+++ views/help-guides/support-representatives.html
@@ -1,25 +1,25 @@
 {% extends "help-guides/support-layout.html" %}
-{% set pageHeader = i18n.helpGuides.representitves.title %}
+{% set pageHeader = i18n.helpGuides.representatives.title %}
 {% block pageTitle %}
-   {{i18n.helpGuides.representitves.title}}   - {{ i18n.general.serviceTitle }} - {{ i18n.general.govUK }}
+   {{i18n.helpGuides.representatives.title}}   - {{ i18n.general.serviceTitle }} - {{ i18n.general.govUK }}
 {% endblock %}
 
 {% block supportTitle %}
-   {{i18n.helpGuides.representitves.title}}
+   {{i18n.helpGuides.representatives.title}}
 {% endblock%}
 
 {% block supportContent %}
-    <p class="govuk-!-font-size-24">{{i18n.helpGuides.representitves.entry}}</p>
-    <p>{{i18n.helpGuides.representitves.what}}</p>
+    <p class="govuk-!-font-size-24">{{i18n.helpGuides.representatives.entry}}</p>
+    <p>{{i18n.helpGuides.representatives.what}}</p>
     <ul class="govuk-list govuk-list--bullet">
-        <li>{{i18n.helpGuides.representitves.whatList.item1}}</li>
-        <li>{{i18n.helpGuides.representitves.whatList.item2}}</li> 
-        <li>{{i18n.helpGuides.representitves.whatList.item3}}</li>
+        <li>{{i18n.helpGuides.representatives.whatList.item1}}</li>
+        <li>{{i18n.helpGuides.representatives.whatList.item2}}</li> 
+        <li>{{i18n.helpGuides.representatives.whatList.item3}}</li>
     </ul>
-    <p>{{i18n.helpGuides.representitves.para1}}</p>
-    <p>{{i18n.helpGuides.representitves.para2 | safe}}</p>
-    <h3 class="govuk-heading-s">{{i18n.helpGuides.representitves.representative.title}}</h3>
-    <p>{{i18n.helpGuides.representitves.representative.para1 | safe}}</p>
-    <p>{{i18n.helpGuides.representitves.representative.para2 | safe}}</p>
-    <p>{{i18n.helpGuides.representitves.representative.para3 | safe}}</p>
+    <p>{{i18n.helpGuides.representatives.para1}}</p>
+    <p>{{i18n.helpGuides.representatives.para2 | safe}}</p>
+    <h3 class="govuk-heading-s">{{i18n.helpGuides.representatives.representative.title}}</h3>
+    <p>{{i18n.helpGuides.representatives.representative.para1 | safe}}</p>
+    <p>{{i18n.helpGuides.representatives.representative.para2 | safe}}</p>
+    <p>{{i18n.helpGuides.representatives.representative.para3 | safe}}</p>
 {% endblock%}
